+++ src/main/java/com/mnan2c/fms/repository/RoleRepository.java
@@ -2,4 +2,4 @@ package com.mnan2c.fms.repository;
 
 import com.mnan2c.fms.entity.Role;
 
-public interface RoleRepository extends BaseRepository<Role, Long> {}
+public interface RoleRepository extends BaseRepository<Role, Integer> {}
